% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{as.data.frame.varpred}
\alias{as.data.frame.varpred}
\title{Coerce varpred object to as.data.frame}
\usage{
\method{as.data.frame}{varpred}(x, ...)
}
\arguments{
\item{x}{varpred object.}

\item{...}{additional aurguments passed to as.data.frame.}
}
\description{
Coerce varpred object to as.data.frame
}
