% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsfuns.R
\name{combinevarpred}
\alias{combinevarpred}
\title{Combine varpred objects}
\usage{
combinevarpred(
  vlist,
  lnames = NULL,
  plotit = FALSE,
  addmarginals = FALSE,
  margindex,
  ...
)
}
\description{
Combines and plots comparison plots for more than two named varpred objects.
}
