% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{data.frame.varpred}
\alias{data.frame.varpred}
\title{Coerce varpred object to data.frame}
\usage{
\method{data.frame}{varpred}(x, ...)
}
\arguments{
\item{x}{varpred object.}

\item{...}{additional aurguments passed to data.frame.}
}
\description{
Coerce varpred object to data.frame
}
