\name{effects-deprecated}
\alias{effects-deprecated}
\alias{all.effects}

\title{Deprecated Function in effects Package}

\description{
The \code{all.effects} function is provided only for compatibility with older versions of the effects package
and may be removed; use \code{\link{allEffects}} instead.
}

\usage{
all.effects(...)
}

\arguments{
  \item{...}{arguments to be passed to \code{allEffects}.}
}

\author{John Fox \email{jfox@mcmaster.ca}.}


\seealso{\code{\link{allEffects}}}
