\name{effectsHexsticker}
\alias{effectsHexsticker}
\title{
View the Official Hex Sticker for the effects Package}
\description{
Open the official hex sticker for the effects package in your browser}
\usage{
effectsHexsticker()
}

\value{
Used for its side effect of openning the hex sticker for the effects package in your browser.
}

\author{
John Fox \email{jfox@mcmaster.ca}
}

\examples{
\dontrun{
effectsHexsticker()
}
}

\keyword{misc}
