% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{getmeans}
\alias{getmeans}
\title{Get focal and prediction means for a varpred object}
\usage{
getmeans(object, what = c("estimate", "focal"), focal = NULL, modelname = NULL)
}
\description{
Get focal and prediction means for a varpred object
}
