% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsfuns.R
\name{getmeans.varpred}
\alias{getmeans.varpred}
\title{Get focal and prediction means for a varpred object}
\usage{
\method{getmeans}{varpred}(object, what = c("estimate", "focal"), focal = NULL, modelname = NULL)
}
\description{
Get focal and prediction means for a varpred object
}
