% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsfuns.R
\name{jdtheme}
\alias{jdtheme}
\title{Customized theme for jdeffects plots}
\usage{
jdtheme()
}
\description{
Sets a theme for jdeffects and other ggplot objects
}
\examples{
library(ggplot2)
jdtheme()
set.seed(4567)
x <- rnorm(100, 3, 5)
y <- 0.4 + 0.7*x + rnorm(100)
df <- data.frame(y = y, x = x)
m1 <- lm(y ~ x, df)

pred1 <- varpred(m1, "x")
plot(pred1)

}
