% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsfuns.R
\name{plot.jdeffects}
\alias{plot.jdeffects}
\title{Plot  predictions}
\usage{
\method{plot}{jdeffects}(x, ..., xlabs = NULL, ylabs = NULL, pos = 0.5)
}
\arguments{
\item{x}{\code{\link[jdeffects]{varpred}} object}

\item{...}{for future implementations}

\item{xlabs}{x-axis label. If \code{NULL}, default, \code{x.var} is used.}

\item{ylabs}{y-axis label. If \code{NULL}, default, the response label is used.}

\item{pos}{spacing between labels of categorical variable on the plot.}
}
\value{
a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plots estimated conditional or marginal predictions.
}
\examples{
set.seed(4567)
x <- rnorm(100, 3, 5)
y <- 0.4 + 0.7*x + rnorm(100)
df <- data.frame(y = y, x = x)
m1 <- lm(y ~ x, df)

pred1 <- varpred(m1, "x")
plot(pred1)

}
\seealso{
\code{\link[jdeffects]{varpred}}
}
