% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodfuns.R
\name{print.varpred}
\alias{print.varpred}
\title{Print varpred objects}
\usage{
\method{print}{varpred}(x, ...)
}
\arguments{
\item{x}{varpred object.}

\item{...}{additional aurguments passed to print.}
}
\description{
Print varpred objects
}
