% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsfuns.R
\name{recoverdata}
\alias{recoverdata}
\title{Recover data from the data from the model}
\usage{
recoverdata(mod, extras = NULL, envir = environment(formula(mod)), ...)
}
\arguments{
\item{mod}{fitted model}

\item{envir}{data environment}

\item{...}{for future implementations}

\item{optional}{character vector or formula specifying the predictors. Important when the transformation are applied in the formula.}
}
\value{
a dataframe
}
\description{
Recover data from the data from the model
}
\details{
It uses the fitted model and the global environment to reconstruct the data used in the model. If \code{data} option is specified in the model formula, a dataframe with columns corresponding to the variable in the formula is returned. Any transformation, e.g. \code{log} specified in the formula terms is not evaluated on the returned data frame. However, if no is provided, the dataframe is constructed from the formula terms with all transformations evaluated.
}
\examples{

set.seed(4567)
x <- rnorm(100, 3, 5)
y <- 0.4 + 0.7*x + rnorm(100)
df <- data.frame(y = y, x = x)
m1 <- lm(y ~ x, df)
d1 <- recoverdata(m1)
head(d1)
m2 <- lm(y ~ x)
d2 <- recoverdata(m2)
head(d2)

}
