% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsfuns.R
\name{zero_vcov}
\alias{zero_vcov}
\title{Zero-out entries of non-focal entries of variance-covariance matrix.}
\usage{
zero_vcov(m, focal_vars, complete)
}
\arguments{
\item{focal_vars}{a character vector specifying the variable(s) appearing on the right side of the formula of the fitted model. The entries of this variable with be non-zero.}

\item{complete}{logical indicating if the full variance-covariance matrix should be returned.}

\item{mod}{fitted model object or a full variance-covariance matrix with terms appearing the same way they appear in the coefficients.}
}
\value{
a variance-covariance matrix of the estimated parameters with non-zero entry corresponding to the focal variable.
}
\description{
Transforms entries of variance-covariance of the coefficients corresponding to the non-focal variables to zero.
}
\examples{
set.seed(4567)
x <- rnorm(100, 3, 5)
y <- 0.4 + 0.7*x + rnorm(100)
df <- data.frame(y = y, x = x)
m1 <- lm(y ~ x, df)
print(zero_vcov(m1, "x"))

}
